Rename http-server -> http, grpc-server -> grpc
