+++ http/server_test.go
@@ -1,4 +1,4 @@
-package httpserver
+package http
 
 import (
 	"context"
@@ -16,7 +16,7 @@ import (
 	"github.com/packethost/pkg/log"
 	"github.com/stretchr/testify/require"
 	"github.com/tinkerbell/hegel/datamodel"
-	grpcserver "github.com/tinkerbell/hegel/grpc-server"
+	"github.com/tinkerbell/hegel/grpc"
 	"github.com/tinkerbell/hegel/hardware"
 	"github.com/tinkerbell/hegel/hardware/mock"
 	_ "github.com/tinkerbell/hegel/metrics" // Initialize metrics.
@@ -892,7 +892,7 @@ func TestServe(t *testing.T) {
 	customEndpoints := `{"/metadata":".metadata.instance"}`
 
 	go func() {
-		if err := Serve(context.Background(), logger, mock.HardwareClient{}, &grpcserver.Server{}, mport, time.Now(), "", customEndpoints, ""); err != nil {
+		if err := Serve(context.Background(), logger, mock.HardwareClient{}, &grpc.Server{}, mport, time.Now(), "", customEndpoints, ""); err != nil {
 			t.Errorf("Serve() error = %v", err)
 		}
 	}()
